Remove types.rs and move its types into the relevant files.
